moving more stuff modularized
